old SingleMicromegasPoolInput renamed to _v1 add SingleMicromegasPoolInput to alias SingleMicromegasPoolInput to SingleMicromegasPoolInput_v2
